Morrowind Update
New provinces: 3 on the Telvanni mainland, 2 on the Telvanni Isles, 1 by
the Thirr, 1 in Jer Ahkrin, 3 in Uvonaak/Kragen Valus, 1 in the mid
Valus, 5 in North Veloth, 1 new wasteland (in the Valus; the setup has
been changed to better reflect where passes are), 4 lakes in
Vvardenfell, 2 lakes elsewhere, and the sea zones south of Elsweyr.
Also, the province map is now a bit less messy in some areas.

Complete with positions. The seemingly random changes to other province
files (e.g. Atron Genkan) are deliberate.